+++ rust/src/problems/p790.rs
@@ -12,7 +12,7 @@ impl Solution {
                 dp[0] = 1;
                 dp[1] = 1;
                 dp[2] = 2;
-                for i in 3..=n as usize {
+                for i in 3..=n {
                     dp[i] = (2 * dp[i - 1] + dp[i - 3]) % MOD;
                 }
                 dp[n] as i32
